RaspberryPi_Code/requirements.txt
opencv-python
ultralytics
pyserial
gpiozero
lgpio
numpy
